chore: moved os helper function in separate helper module
